Move to new class.
